Add Basic Websockets Infrastructure
Add a new Sockets module containing some basic WebSockets code. It
defines an API handler that handles incoming websocket connections and
returns a 400 for normal API requests.

The connection handler accepts an incoming connection, determines it's
UserId, maintains a mapping from User IDs to their connections, and
simply prints out any received messages. When the client disconnects, it
remove the UserId from it's map. The mapping can be used throughout the
codebase(but currently isn't) to send a message to a specific client.

Expose a new `/websockets` route in the HttpServer app for websocket
clients to connect to.