Add "Chat" WebSocket Support, Refactor WebSocket Infrastructure
Rework a bunch of the websockets related code and move them into
discrete modules.

A new `Sockets.Connections` module is used for creating, manipulating,
querying, & utilizing the `ConnectionMap` type while hiding the internal
representation.

The new `Sockets.Messages` module adds support for designating types as
Sendable/Receivable websocket messages as well as functions for
sending/broadcasting messages, creating handlers, and executing handlers
when we receive messages.

The existing `Sockets` module now contains only the hook-up code for
running a websocket service under the API server. We expand the runner
code to take an arbitrary list of message handlers. This list is fixed
at this point in time, but should be moved into the ConnectionMap so
that handlers can be dynamically added & removed through the lifetime of
the server.

A new `Chat` module has been added to explore an ultra-minimal usage of
this new architecture. The websockets portion of this allows receiving
new messages from clients & informing clients about new messages. The
redis portion of this is responsible for ensuring that _all_ clients
connected across all servers receive newly posted messages.

Our plan for future explorations is to allow clients to subscribe to a
subset of all potential messages. E.g., I should only get messages for
chatrooms that I am in, instead of every message that every server
receives.