Refactor ISCB & Redis Ping Messages, Add WebSocket Message
Refactor the ISCB module so that we no longer need to unify all
potential channels in a single type. Allow channel values to have
parameters and restructure the message to channel conversion so the same
message types can be sent to different channels depending on the data
they hold.

Rework the MessageToChannel typeclass, transforming it into a type
family that specifies it's channel data type as well as the encoding of
that type into a RedisChannel. Modify the `makeMessageHandler` function
so that it takes a channel of the correct type as it's message. This
allows us to map multiple channels to the same message type.

Move all the "ping" code from the `ISCB` module into it's own `Ping`
module, adding a new websocket message that is broadcast to all clients
when the `/ping1` & `/ping2` endpoints are hit.