Use UUIDs as UserIds
We want to explore how clients might subscribe to a subset of potential
messages instead of only being able to receive messages directly for
them or for everybody.

This requires unique client IDs across all servers. Normally this would
be handled by our UserId database key, but we use UUIDs in this
prototype.