Add Support for WebSocket Subscriptions
Add a new `Sockets.Subscriptions` module for managing client
subscriptions to subsets of websocket messages. The
`WebsocketSubscription` type family is used to mark a type as
subscribable as well as identify related types & allowing hooking in to
various stages of subscription management.

Consumers would create 3 types, one for the subscription, one for
receivable messages, & one for sendable messages. Handlers with implicit
subscription management can be created with `makeSubscriptionHandler`
and messages can be sent to all subscribed clients with
`sendMessageToSubscription`.

Expand the `WebsocketController` type to include a mapping of
subscriptions to subscribed users. Add new functions for subscribing &
unsubscribing a client, which do the additional work of
subscribing/unsubscribing to their associated redis message passthrough
channels when appropriate.

Add a new `ChatRooms` module which demonstrates how to use the new
`Sockets.Subscriptions` module. This is similar to the `Chat` module but
requires users to subscribe to a specific room before they receive that
room's messages.